+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix uuidv4 import
 - Fix missing uuidv4 import in integration test
 - Disable w3c check for Chrome Android webdriver integration tests
+- Fix setSinkId() from throwing DOMException in Chromium browsers
 
 ## [1.10.0] - 2020-06-23